extract private methods to trait
